v0.7.1: パート簡易編集・データ保存読出し・プリセット管理刷新
機能追加:
- パート簡易編集メニュー（演奏中でも即座にパート編集可能）
- パート左右バランス（Pan）設定（CC#10、11段階）
- アルペジオパターン単体のJSON保存/読出し機能
- コード進行データ単体のJSON保存/読出し機能
- コード変更時のアルペジオ動作設定（Restart/Continue）
- アルペジオ編集UIのstep_per_beat連動（動的ページサイズ・ビート番号表示）

データ管理:
- incbinプリセットをカテゴリ別に分割（song_genre/song_song）
- storage_incbin_tのパラメータ化・X-macroパターン導入
- JSONキー名 "sequence" → "progression" に変更（下位互換フォールバック付き）
- JSON保存/読出し共通関数の抽出（ソング・アルペジオ・コード進行で共用）

コマンド体系:
- part_editをpart_edit_menu/part_edit_enterに分離
- part_edit_enterにパート番号パラメータ対応

名称リファクタ:
- sequence/seqmode系 → progression/playmode系に統一
- reg_sequence_timeline_t → reg_chord_progression_t
- ui_sequence_timeline_t → ui_progression_timeline_t
- プリセットJSONから未使用play_modeキーを削除

メニュー構成:
- パート簡易編集にCopy/Paste・Clear All Notes・パターン読込を追加
- ソング保存/読込をSong/Chord Progressionに分割
- プリセットソングをPreset Genre/Preset Songに分割
- ファイル読み書きメニューをmenu_data_file.inlに分離

改善・修正:
- OTAリダイレクト時のContent-Type検証追加
- コンパイル警告修正（reserve typo・override・printf format）
- dir_manage配列のインデックスをenum名で参照
- data_arpeggio_user → data_arpeggio_users 命名統一

+++ main/main.cpp
@@ -63,8 +63,8 @@ namespace kanplay_ns {
   #endif
 
     M5.Display.setTextSize(2);
-    M5.Display.printf("KANTAN Play\nver%lu.%lu.%lu\n\nboot"
-      , def::app::app_version_major, def::app::app_version_minor, def::app::app_version_patch);
+    M5.Display.printf("KANTAN Play\nver%d.%d.%d\n\nboot"
+      , (int)def::app::app_version_major, (int)def::app::app_version_minor, (int)def::app::app_version_patch);
 
     {
       static constexpr const uint8_t aw9523_i2c_addr = 0x58;
